chore(deriver): delete the two legs measured negative in August
`DERIVER_DATE_AUDIT`: 73.0 against 77.8 with p=0.0008 and temporal
−10.3pp (p=0.004). The verdict written next to it in v11 §13 is
"NEGATIVE, retire" — session-date defaults ARE the benchmark's answer
convention and the audit strips them. It was never retired.

`DERIVER_ASPECT_ROLLUPS`: 76.0 (p=0.23) with multi-hop −6.3 the wrong
way. "NULL/neg — A2 does not transfer; rollups displace atoms."

Both carried a flag, a code path, an extra LLM call per session, a
catalogue row an operator had to read, and a place in every future
"which flags do we still have" count. Gone: the prompt, the pass, the
profile fields, the env names, the catalogue entries, the tests.

What SURVIVES, deliberately: `aspect-rollups.ts` held two things — the
rollup composer and the landed-row pool it accumulated over. The pool
is shared with DERIVER_COMPOSE_PASS, a separate leg that has never been
measured, so only the composer goes. The file is renamed `derive-pool.ts`
to say what it now is, rather than keeping a name that describes the
deleted half.

`DeriveRunResult.rollups` goes with it: nothing wrote that counter any
more, and a result field nobody fills is a lie waiting for a reader.

-552/+37 lines. Unit 5707, full e2e 856/167 suites green.
