+++ src/admin/deriver-client.ts
@@ -117,127 +117,6 @@ GROUNDING QUOTES
 For every proposition also fill "quotes" — an array PARALLEL to "turns": for each grounding turn, the shortest VERBATIM snippet from that turn's text that supports the proposition, copied character-for-character (never paraphrased, never spanning multiple turns, never including the [N] / speaker prefix). Use null for a turn with no single verbatim supporting snippet. "quotes" must have exactly as many entries as "turns".`;
 
 /**
- * V13 date audit (DERIVER_DATE_AUDIT) — the post-pass shape of the
- * failed in-prompt date rules. Measured lineage: prose rules inside
- * the extraction prompt moved NOTHING (wd-v4 date distribution
- * byte-equal to the replicates, armH null), while the same contract as
- * a dedicated after-emission turn is exactly how salience grading
- * succeeded after ITS in-prompt version failed both gates (V8 §4 →
- * V9 §5). One extra cheap call per session; failure degrades to the
- * un-audited dates.
- */
-export const DATE_AUDIT_SYSTEM = `You audit the event dates of memory propositions extracted from ONE dated dialogue session. For every numbered proposition decide "occurred_on" — the ISO date (YYYY-MM-DD) the described EVENT actually happened:
-- Resolve relative expressions from the transcript ("yesterday", "last Friday", "two weekends ago", "next month") by calendar arithmetic from the session date.
-- The session date is correct ONLY for events that happened during that same day ("today I…", "this morning").
-- Planned or future events take the planned occurrence date when stated or derivable.
-- Month-only knowledge resolves to the FIRST day of that month; year-only to January 1st.
-- Genuinely undeterminable: null. A wrong default is worse than no date.
-Return a decision for EVERY index. Output strictly the JSON schema.`;
-
-const ISO_DAY_RE = /^\d{4}-\d{2}-\d{2}$/;
-
-async function auditDates(
-  deps: DeriverClientDeps,
-  args: {
-    sessionDate: Date;
-    /** The base deriver conversation — continued, not re-prefixed, so
-     *  the transcript (the dominant token mass) stays a prompt-cache
-     *  hit like completionUnion's continuation 30 lines up. */
-    messages: ChatMessage[];
-    propositions: DerivedProposition[];
-  },
-): Promise<void> {
-  try {
-    const list = args.propositions
-      .map(
-        (p, i) =>
-          `${i}. [${p.subject}] ${p.proposition} (current occurred_on: ${p.occurred_on ?? 'null'})`,
-      )
-      .join('\n');
-    const res = await deps.openai.chat.completions.create({
-      model: deps.model,
-      ...chatCallParams(deps.model, { temperature: 0, visibleCap: 4000 }),
-      messages: [
-        ...args.messages,
-        { role: 'assistant', content: `Propositions:\n${list}` },
-        {
-          role: 'user',
-          content: `${DATE_AUDIT_SYSTEM}\n\nSession date: ${args.sessionDate.toISOString().slice(0, 10)}. Audit the numbered propositions above.`,
-        },
-      ],
-      response_format: {
-        type: 'json_schema',
-        json_schema: {
-          name: 'date_audit',
-          strict: true,
-          schema: {
-            type: 'object',
-            additionalProperties: false,
-            properties: {
-              dates: {
-                type: 'array',
-                items: {
-                  type: 'object',
-                  additionalProperties: false,
-                  properties: {
-                    index: { type: 'integer' },
-                    occurred_on: { type: ['string', 'null'] },
-                  },
-                  required: ['index', 'occurred_on'],
-                },
-              },
-            },
-            required: ['dates'],
-          },
-        },
-      },
-    });
-    const content = res.choices[0]?.message?.content;
-    if (!content) throw new Error('empty date-audit response');
-    const parsed = JSON.parse(content) as {
-      dates?: Array<{ index: number; occurred_on: string | null }>;
-    };
-    const dates = parsed.dates ?? [];
-    // An index === length entry is the perfect tell of a 1-based reply:
-    // applying a shifted map would stamp every proposition with its
-    // NEIGHBOR'S date — strictly worse than the un-audited baseline the
-    // degrade contract promises. Abort the whole application.
-    if (dates.some((d) => d.index === args.propositions.length)) {
-      deps.logger.warn('date audit reply looks 1-based (index === length) — discarding the audit');
-      return;
-    }
-    for (const d of dates) {
-      if (!Number.isInteger(d.index) || d.index < 0 || d.index >= args.propositions.length) {
-        continue;
-      }
-      // d.index validated in-bounds by the guard above ⇒ prop is present.
-      const prop = args.propositions[d.index]!;
-      // Apply BOTH values and explicit nulls — clearing a fabricated
-      // session-date default is the audit's whole point. The cleared
-      // marker survives to the row builder, which renders "undated" as
-      // the epoch sentinel instead of the session-date fallback.
-      if (d.occurred_on === null) {
-        prop.occurred_on = null;
-        prop.dateCleared = true;
-      } else if (
-        typeof d.occurred_on === 'string' &&
-        ISO_DAY_RE.test(d.occurred_on) &&
-        // Calendar round-trip — the same guard the row builder uses:
-        // shape-only acceptance let an impossible date ('2023-02-30')
-        // overwrite a CORRECT one and then silently collapse to the
-        // session date downstream.
-        new Date(`${d.occurred_on}T00:00:00.000Z`).toISOString().slice(0, 10) === d.occurred_on
-      ) {
-        prop.occurred_on = d.occurred_on;
-        prop.dateCleared = false;
-      }
-    }
-  } catch (e) {
-    deps.logger.warn(`date audit turn failed (${(e as Error).message}); dates stay un-audited`);
-  }
-}
-
-/**
  * Salience grading (DERIVER_SALIENCE_STAMP, V8 §4 → V9 §5 rebuild):
  * a SEPARATE cheap turn over the emitted proposition list. The V8
  * in-prompt section failed both its gates — it primed over-emission
@@ -395,13 +274,6 @@ export async function callDeriver(
   ];
   const base = await deriverPass(deps, messages);
   const merged = await completionUnion(deps, { profile, messages, base });
-  if (profile.deriveDateAudit && merged.length > 0) {
-    await auditDates(deps, {
-      sessionDate: args.sessionDate,
-      messages,
-      propositions: merged,
-    });
-  }
   if (profile.deriveSalienceStamp && merged.length > 0) {
     await gradeSalience(deps, merged);
   }
